+++ src/User.php
@@ -5361,7 +5361,6 @@ HTML;
         NotificationEvent::raiseEvent($event, $this, [
             'entities_id' => $entities_id,
         ]);
-        QueuedNotification::forceSendFor($this->getType(), $this->fields['id']);
 
         return true;
     }